# app/api/spotify.py
import base64
import logging
import os
import uuid
from datetime import UTC, datetime, timedelta
from urllib.parse import urlencode

import httpx
from fastapi import APIRouter, Depends, HTTPException, Query, Request, Response
from fastapi.responses import RedirectResponse
from sqlalchemy.ext.asyncio import AsyncSession

from app.api.deps import get_current_user
from app.api.validation import (
    ensure_exists,
    raise_http_error,
    raise_unauthorized,
    raise_validation_error,
)
from app.auth.security import AccessTokenConfig, create_access_token, decode_token
from app.core.circuit_breaker import (
    CircuitBreaker,
    CircuitBreakerConfig,
    CircuitBreakerOpenError,
)
from app.core.config import settings
from app.core.database import get_db
from app.core.localization import resolve_locale, translate
from app.models.models import SpotifyIntegration, User
from app.schemas.schemas import SpotifyAuthURL, SpotifyNowPlayingOut

logger = logging.getLogger(__name__)

router = APIRouter(prefix="/spotify", tags=["spotify"])

# Circuit breaker for Spotify API with sensible defaults
_spotify_circuit_breaker = CircuitBreaker(
    "spotify_api",
    config=CircuitBreakerConfig(
        failure_threshold=5,
        recovery_timeout_seconds=30.0,
        success_threshold=2,
        excluded_exceptions=(HTTPException,),  # Don't count app-level errors
    ),
)


def _now_utc() -> datetime:
    return datetime.now(UTC)


def _ensure_utc(dt: datetime | None) -> datetime | None:
    if dt is None:
        return None
    return dt.astimezone(UTC) if dt.tzinfo else dt.replace(tzinfo=UTC)


def _b64(s: str) -> str:
    return base64.b64encode(s.encode()).decode()


def _coerce_expires(value: int | str | None) -> int:
    try:
        seconds = int(value) if value is not None else 3600
    except (TypeError, ValueError):
        seconds = 3600
    # Spotify may technically return small values during testing; make sure
    # we always refresh slightly before the real expiry to avoid rapid loops.
    return max(seconds, 30)


def _disconnect_user(
    user: User, *, clear_refresh: bool = False, clear_profile: bool = False
) -> None:
    if not user.spotify:
        return
    user.spotify.access_token = None
    if clear_refresh:
        user.spotify.refresh_token = None
        user.spotify.scope = None
    user.spotify.token_expires_at = None
    user.spotify.is_connected = False
    user.spotify.is_playing = False
    if clear_profile:
        user.spotify.display_name = None
        user.spotify.spotify_user_id = None


def _fallback_now_playing(user: User) -> SpotifyNowPlayingOut:
    if not user.spotify:
        return SpotifyNowPlayingOut(is_playing=False, fetched_at=_now_utc())

    artists: list[str] = []
    if user.spotify.last_artist_name:
        artists = [
            name.strip()
            for name in user.spotify.last_artist_name.split(",")
            if name.strip()
        ]

    has_payload = any(
        (
            user.spotify.last_track_id,
            user.spotify.last_track_name,
            artists,
        )
    )

    if not has_payload:
        return SpotifyNowPlayingOut(is_playing=False, fetched_at=_now_utc())

    return SpotifyNowPlayingOut(
        is_playing=False,
        progress_ms=None,
        duration_ms=None,
        track_id=user.spotify.last_track_id,
        track_name=user.spotify.last_track_name,
        artists=artists,
        album_name=user.spotify.last_album_name,
        album_image_url=user.spotify.last_album_image_url,
        track_url=user.spotify.last_track_url,
        preview_url=None,
        fetched_at=_now_utc(),
    )


async def _save_tokens(
    db: AsyncSession,
    user: User,
    access: str,
    refresh: str | None,
    scope: str | None,
    expires_in: int | str | None,
):
    if not user.spotify:
        user.spotify = SpotifyIntegration(user_id=user.id)

    user.spotify.access_token = access or None
    if refresh is not None:
        user.spotify.refresh_token = refresh or None
    seconds = _coerce_expires(expires_in)
    # Refresh a little earlier to compensate for latency and clock skew.
    user.spotify.token_expires_at = _now_utc() + timedelta(seconds=seconds - 10)
    user.spotify.scope = scope or ""
    user.spotify.is_connected = True
    await db.commit()
    await db.refresh(user)


async def _ensure_access_token(
    db: AsyncSession, user: User, *, locale: str | None = None
) -> str | None:
    """Return a usable Spotify access token, refreshing it when possible.

    Previously we returned ``None`` when ``spotify_access_token`` was empty,
    even if a refresh token was still stored. After the backend was restarted,
    some users ended up in that exact state and the frontend kept showing the
    stale fallback data until they manually reconnected the integration. To
    avoid forcing a re-link, we now treat a missing access token the same way
    as an expired one and trigger the refresh flow when a refresh token exists.
    """

    now = _now_utc()
    if not user.spotify:
        return None
    token = user.spotify.access_token or None
    exp = _ensure_utc(user.spotify.token_expires_at)
    if token and exp and exp > now:
        return token
    refresh_token = user.spotify.refresh_token or None
    if not refresh_token:
        if not token and not user.spotify.is_connected:
            # The user never connected Spotify or already disconnected it;
            # returning ``None`` allows the caller to fall back gracefully.
            return None
        _disconnect_user(user, clear_refresh=True)
        await db.commit()
        raise_unauthorized(locale, "errors.spotify.reconnect_required")
    try:
        async with _spotify_circuit_breaker:
            async with httpx.AsyncClient(timeout=15) as client:
                r = await client.post(
                    "https://accounts.spotify.com/api/token",
                    data={
                        "grant_type": "refresh_token",
                        "refresh_token": refresh_token,
                    },
                    headers={
                        "Authorization": "Basic "
                        + _b64(
                            f"{settings.spotify_client_id}:{settings.spotify_client_secret}"
                        )
                    },
                )
    except CircuitBreakerOpenError as exc:
        logger.warning(
            "Spotify circuit breaker open, using fallback",
            extra={"remaining_seconds": exc.remaining_seconds},
        )
        return None
    if r.status_code != 200:
        _disconnect_user(user, clear_refresh=True)
        await db.commit()
        raise_unauthorized(locale, "errors.spotify.reconnect_required")
    data = r.json()
    access_token = data.get("access_token")
    if not access_token:
        _disconnect_user(user, clear_refresh=True)
        await db.commit()
        raise_unauthorized(locale, "errors.spotify.reconnect_required")
    await _save_tokens(
        db,
        user,
        access_token,
        data.get("refresh_token"),
        data.get("scope"),
        data.get("expires_in"),
    )
    return user.spotify.access_token


@router.get("/auth-url", response_model=SpotifyAuthURL)
async def spotify_auth_url(user: User = Depends(get_current_user)):
    state_result = await create_access_token(
        str(user.id), config=AccessTokenConfig(expires_delta=10)
    )
    state_token = state_result[0] if isinstance(state_result, tuple) else state_result
    logger.error(
        f"DEBUG: spotify_client_id from settings: '{settings.spotify_client_id}'"
    )
    logger.error(
        f"DEBUG: env var SPOTIFY_CLIENT_ID: '{os.environ.get('SPOTIFY_CLIENT_ID')}'"
    )
    params = {
        "client_id": settings.spotify_client_id,
        "response_type": "code",
        "redirect_uri": settings.spotify_redirect_uri,
        "scope": settings.spotify_scopes,
        "state": state_token,
        "show_dialog": "false",
    }
    url = "https://accounts.spotify.com/authorize?" + urlencode(params)
    logger.error(f"DEBUG: Generated URL: {url}")
    return {"url": url}


@router.get("/callback")
async def spotify_callback(
    request: Request,
    code: str = Query(...),
    state: str = Query(...),
    db: AsyncSession = Depends(get_db),
):
    locale = resolve_locale(request=request)
    payload = decode_token(state) or {}
    if not payload.get("sub"):
        raise_validation_error("errors.spotify.invalid_state", locale)
    user = await db.get(User, uuid.UUID(payload["sub"]))
    ensure_exists(user, "spotify.user_not_found", locale)
    locale = resolve_locale(request=request, user=user)
    try:
        async with _spotify_circuit_breaker:
            async with httpx.AsyncClient(timeout=15) as client:
                r = await client.post(
                    "https://accounts.spotify.com/api/token",
                    data={
                        "grant_type": "authorization_code",
                        "code": code,
                        "redirect_uri": settings.spotify_redirect_uri,
                    },
                    headers={
                        "Authorization": "Basic "
                        + _b64(
                            f"{settings.spotify_client_id}:{settings.spotify_client_secret}"
                        )
                    },
                )
    except CircuitBreakerOpenError:
        raise_http_error(503, "errors.spotify.service_unavailable", locale)
    if r.status_code != 200:
        raise_validation_error("errors.spotify.token_exchange_failed", locale)
    data = r.json()
    await _save_tokens(
        db,
        user,
        data["access_token"],
        data.get("refresh_token"),
        data.get("scope"),
        int(data.get("expires_in", 3600)),
    )
    try:
        async with _spotify_circuit_breaker:
            async with httpx.AsyncClient(timeout=15) as client:
                me = await client.get(
                    "https://api.spotify.com/v1/me",
                    headers={"Authorization": f"Bearer {user.spotify.access_token}"},
                )
    except CircuitBreakerOpenError:
        me = None
    if me is not None and me.status_code == 200:
        info = me.json()
        user.spotify.spotify_user_id = info.get("id") or None
        user.spotify.display_name = info.get("display_name") or None
        await db.commit()
    # Redirect to the frontend, not the backend
    frontend_base = settings.frontend_origin.rstrip("/")
    target = f"{frontend_base}/profile?spotify=connected"
    return RedirectResponse(target, status_code=302)


@router.get("/now-playing", response_model=SpotifyNowPlayingOut)
async def now_playing(
    request: Request,
    db: AsyncSession = Depends(get_db),
    user: User = Depends(get_current_user),
):
    locale = resolve_locale(request=request, user=user)

    def _as_response(data: SpotifyNowPlayingOut):
        if data.track_id or data.track_name or data.artists:
            return data
        return Response(status_code=204)

    async def _request(access_token: str) -> httpx.Response | None:
        try:
            async with _spotify_circuit_breaker:
                async with httpx.AsyncClient(timeout=15) as client:
                    return await client.get(
                        "https://api.spotify.com/v1/me/player/currently-playing",
                        headers={"Authorization": f"Bearer {access_token}"},
                    )
        except CircuitBreakerOpenError:
            return None

    token = await _ensure_access_token(db, user, locale=locale)
    if not token:
        return _as_response(_fallback_now_playing(user))

    r = await _request(token)
    if r is None:
        # Circuit breaker is open, return fallback data
        return _as_response(_fallback_now_playing(user))
    if r.status_code == 401:
        # Access tokens occasionally expire slightly earlier than advertised or
        # get invalidated server-side. Instead of forcing the user to reconnect
        # immediately, attempt a single refresh and retry before giving up.
        user.spotify.access_token = None
        user.spotify.token_expires_at = _now_utc() - timedelta(seconds=30)
        try:
            refreshed = await _ensure_access_token(db, user, locale=locale)
        except HTTPException:
            # ``_ensure_access_token`` already disconnected the user and raised
            # an appropriate error; propagate it unchanged.
            raise
        if not refreshed:
            return _as_response(_fallback_now_playing(user))
        r = await _request(refreshed)
        if r is None:
            return _as_response(_fallback_now_playing(user))
        if r.status_code == 401:
            _disconnect_user(user, clear_refresh=True)
            await db.commit()
            raise_unauthorized(locale, "errors.spotify.reconnect_required")
    if r.status_code == 204:
        user.spotify.is_playing = False
        user.spotify.last_checked_at = _now_utc()
        await db.commit()
        return Response(status_code=204)
    if r.status_code == 401:
        _disconnect_user(user, clear_refresh=True)
        await db.commit()
        raise_unauthorized(locale, "errors.spotify.reconnect_required")
    if r.status_code == 429:
        retry_after_header = r.headers.get("Retry-After")
        try:
            retry_after = (
                max(1, int(float(retry_after_header))) if retry_after_header else 5
            )
        except (TypeError, ValueError):
            retry_after = 5
        user.spotify.last_checked_at = _now_utc()
        await db.commit()
        raise HTTPException(
            status_code=429,
            detail=translate("errors.spotify.rate_limited", locale=locale),
            headers={"Retry-After": str(retry_after)},
        )
    if r.status_code != 200:
        return _as_response(_fallback_now_playing(user))
    j = r.json()
    item = j.get("item") or {}
    artists = [a.get("name") for a in item.get("artists", []) if a.get("name")]
    album = item.get("album") or {}
    images = album.get("images") or []
    img = images[0]["url"] if images else None
    url = item.get("external_urls", {}).get("spotify")
    preview = item.get("preview_url")
    user.spotify.is_playing = bool(j.get("is_playing"))
    user.spotify.last_checked_at = _now_utc()
    user.spotify.last_track_id = item.get("id")
    user.spotify.last_track_name = item.get("name")
    user.spotify.last_artist_name = ", ".join(artists) if artists else None
    user.spotify.last_album_name = album.get("name")
    user.spotify.last_track_url = url
    user.spotify.last_album_image_url = img
    await db.commit()
    return SpotifyNowPlayingOut(
        is_playing=bool(j.get("is_playing")),
        progress_ms=j.get("progress_ms"),
        duration_ms=(item.get("duration_ms") if item else None),
        track_id=item.get("id"),
        track_name=item.get("name"),
        artists=artists,
        album_name=album.get("name"),
        album_image_url=img,
        track_url=url,
        preview_url=preview,
        fetched_at=_now_utc(),
    )


@router.post("/disconnect")
async def disconnect(
    db: AsyncSession = Depends(get_db), user: User = Depends(get_current_user)
):
    _disconnect_user(user, clear_refresh=True, clear_profile=True)
    await db.commit()
    return {"ok": True}


@router.get("/playlists")
async def list_playlists(
    request: Request,
    db: AsyncSession = Depends(get_db),
    user: User = Depends(get_current_user),
):
    locale = resolve_locale(request=request, user=user)
    token = await _ensure_access_token(db, user, locale=locale)
    if not token:
        raise_unauthorized(locale, "errors.spotify.reconnect_required")

    async with httpx.AsyncClient(timeout=15) as client:
        r = await client.get(
            "https://api.spotify.com/v1/me/playlists",
            headers={"Authorization": f"Bearer {token}"},
        )
        if r.status_code != 200:
            raise_http_error(r.status_code, "errors.spotify.api_error", locale)
        return r.json()


@router.post("/sync-playlists")
async def sync_playlists(
    request: Request,
    db: AsyncSession = Depends(get_db),
    user: User = Depends(get_current_user),
):
    """
    Synchronize Spotify playlists metadata to the local database.
    """
    # Placeholder implementation
    return {"status": "success", "synced_count": 0}